+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Next
+# v3.1.0
 
 ### ✨ Features